requirements.txt
pandas==2.2.3
openpyxl==3.1.2
fastapi
uvicorn==0.21.1
google-cloud-aiplatform==1.36.2
pydantic
python-multipart==0.0.6
google-generativeai==0.3.1
# google-genai
google-auth==2.28.0
python-dotenv==1.0.1
pytest==8.3.5
pytest-asyncio==1.0.0
httpx==0.28.1
streamlit==1.28.0
google-cloud-storage==2.9.0
python-multipart==0.0.6
vertexai